+++ CHANGELOG.md
@@ -13,6 +13,7 @@
 ### Added
 * Tour guide for new features (`OSIDB-4583`)
 * Allow grouping affects by ps_module (`OSIDB-4546`)
+* Offer Impact suggestions from Aegis (`AEGIS-224`)
 
 
 ## [2025.10.0]